requirements.txt
feedparser
praw
python-dotenv
python-telegram-bot
google-generativeai
